pyproject.toml
[tool.poetry]
name = "web-hw-14"
version = "0.1.0"
description = ""
authors = ["Olha Lialina <145963563+olhalialina@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
fastapi = {extras = ["all"], version = "^0.110.3"}
psycopg2-binary = "^2.9.9"
python-jose = "^3.3.0"
libgravatar = "^1.0.4"
cloudinary = "^1.40.0"
python-multipart = "^0.0.9"
fastapi-limiter = "^0.1.6"
redis = "^5.0.4"
fastapi-mail = "^1.4.1"
pyjwt = "^2.8.0"
jwt = "^1.3.1"
uvicorn = {extras = ["standart"], version = "^0.29.0"}
pydantic = {extras = ["email"], version = "^2.7.1"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}





[tool.poetry.group.dev.dependencies]
sphinx = "^7.3.7"


[tool.poetry.group.test.dependencies]
aiosqlite = "^0.20.0"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.6"
pytest = "^8.2.0"
pyjwt = "^2.8.0"
jwt = "^1.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = [
    "tests", ]
pythonpath = "."
filterwarnings = "ignore::DeprecationWarning"